Don't test nodenv-alias, but just the hooks
Leverage bats-mock to mock out nodenv-alias itself. We only need to test
that the hook invokes nodenv-alias correctly. Otherwise we are
redundantly testing nodenv-aliases behavior (and that means we need all
of the setup and assertions necassary for that). Rather, we can just
test that the hooks invoke nodenv-alias as expected and test the other
behaviors of the hook in isolation.