Don't rely on delayed hook eval
Since nodenv's hook functions only save the callback functions' names,
it presently allows the callbacks to be defined *after* we register
them. However, it is conceivable that nodenv changes it behavior to
invoke them immediately rather than deferring. We shouldn't rely on its
current behavior and should instead ensure that the callback functions
are defined before we register them.